requirements.txt
streamlit
google-genai
python-dotenv
requests
Pillow
schedule
imageio
imageio-ffmpeg
numpy
beautifulsoup4
